Update gulp to 4.0.2.
Update gulp and resolve issues with the gulp 3 > 4 migration.
Renamed some anonymous functions to make the errors more clear in Gulp when they occur.
Note that the Gulp file may not be optimal as far as build times (using series instead of parallel), but this shouldn't matter since the build is fast enough as-is.
vinyl-source-stream must be updated to 2.0.0 as part of the migration to Gulp 4.

This update to Gulp is required because of an issue present in Gulp 3.9.1 and Node.js > 11.5 where a graceful-fs error was preventing Gulp from running.